requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
python-dotenv==1.0.0
Authlib==1.2.1
requests==2.31.0